added also option to flip negative average ecg epochs.
Flipping: find min and max of data. If min of data is higher in magnitude then max -flip.
In most cases it will flip correctly, but if the data. is quite flat and doesnt have a good ECG peak, it can still be flipped even when not disired. Should not influence the final result much though. Because flat ecg epoch in the end will anyways be sorted out as not affected by artifact. However flipping will always influence the average. It is not the real average now. However the benefit is: the R wave can not be cancelled out in averaging. if you flipped. all channels to. positive. (and it might be cancelled out if we got same amount of positive and negative epochs without flipping).