trying all existing data sets with eog. still dont understand how the make epochs when several eog channels present
I see that epochs in case of 2 channels created on base of only 1 channel, but how is it chosen? WE only got 2 data sets with 2 eog channels: 1 from open neuro (with 1 bad,  1 good eog) and on. from lab - forrest gump. In both cases epochs created just on 2 channel, even though in case of forrest gump second channel looks fine as well. No explanation in tutorials or on mne forum.