Again new approach to flip: if the peak is negative and it s closer than positive to time 0 of the event - flip.
In other cases, the "negative comes first' condition is not always good, cos sometimes the positive comes first if the time window is big enough. Taking a small time frame - the actual R wave is not fought in some sets, because the wave's peak is again moved in relation to time 0 of the event. The reason for many problems can be the unperfect detection of the time 0 of the event by mne, which my algorithm relies on. Rewriting event detection could take too long and I don't know how to start.