adding note about grads to muscle detection and report
Magnetometers are more sensitive to muscle artifacts and are recommended for artifact detection. But is it accurate on the base of grads or should the metric be skipped on the base of grads? Tried and compared 7 .fif files calculating for both mags and grads: times are the same, but when using grads some muscle events are just skipped.
So use grads still. But give a note:
Magnetometers are more sensitive to muscle artifacts and are recommended for artifact detection. If you only use gradiometers, some muscle events might not show. This will not be a problem if the data set only contains gradiometers. But if it contains both gradiometers and magnetometers, but only gradiometers were chosen for this analysis - the results will not include an extra part of the muscle events present in magnetometers data.