moving detection of noisy ecg/eog channels inside of their modules + correcting reports and jsons
Maybe not need to use noisy channel detection and in any case, calculate the artifact and already in the end say if it was successful or not? Check if reconstruction does better than noisy channels! The problem is: need to choose peak detection for noisiness very carefully or it will get a false positive or false negative 'bad' channel.