trying to drop vs keep bad ecg channel
Tried 2 data sets (total 7 fifs from openneuro. (dont have more) which have BAD ECG channels. In all cases reconstruction is same as bad as keeping the bad channel (different, but bad). Maybe data was preprocessed already and they dont have ECG artifacts? Still no decision if it s better to drop or not. Made this as a setting in config. With default to drop.