adding plot of the whole time series with plotly to report
The data for it is resampled to 100Hz. maybe need even smaller. This resampling and plotting takes significant amout of time (for example about 1 min for 2 2 min data file on my mac) + this plot in  html report is very slow. plot can be set to true/false in config.
much more useful than mne plot for inspecting noisy channels.