converting channels into objects
temporal commit: here var channels is partially renamed into channels_objs, which means it hold not only a dictionary with ch names but a dict with entire ch objects. Each obj has with a bunch of info about the channel.
Next commit I ll change it back to simple channels as names, because usually I only use names all over the code and it ll take too long to change handling of this var now.
However, in the future if desired it can be changed to an object everywhere and all the info about noisines, etc for each channel could be added inside of the object.