Making channels in psd and time series better visible still sorted by lobe color
Now the channels are not plotted grouped by color righ away. Their plotly traces are created. Then they are sorted in random order and plotted. Then only the legend is sorted and updated. This way different color lay on top of each other for better visibility. 8 different colors are used.
Plotly default (if no color tagging would be used ) would offer 10 colors, which is not a big difference.