adding gaussian filter (gaussian kernel convolution) for detecting ECG EOG wave
now results are much better. Kept both with and without gaussian versions. Can adjust the settings and maybe keep only the version with gaussian.

1) Right now both versions: with and without gaussian can be calculated for every channel + for average over all.

2) Then, given the number of peaks on the AVERAGE - the calculation of affected and non affected channels will be done or skipped. Number of peaks calculated  both with and without filter. If any of them says artifact is ok - continue.

3) Then detect affected channels. However! Affected channels used here are only non gaussian smoothed. Just dont look at the smoothed ones. Advantage: we see the actual data of artifact. Disadvantage: often data is noisy and may show no wave shape where it actually is present - results to false negatives.

So next, I would try to use for the step 3 also the smoothed data of all channels and see if performance is better and worth the loss of data.