Adding correlation with reconstructed
Adding a check for reconstructed or recorded MEAN Rwave: is it a wave or not. if not - skip further calculation.
Fixing bugs in correlation algorythm.

Now only the correlation method is used. Note that since thresholding over mean is not developed at the moment, it might not work properly from now on.

Sill problems:
- sometimes mean reconstructed is still recognised as wave even when it might be not  proper good R wave.
Then we correlate with it - might get false positive.
- sometimes correlation coefficient is higher not for the channels with the highest R peak, but for the channels with pretty law R peak but overall better correlating wave shape. In this case sorting by most affected channels is not really accurate.
To see that run ds003483 for all subjects. some will have listed issues.